+++ CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - `CpxAp.read_global_diagnosis_state()` now returns dict according to "module diagnostics state" (see datasheet CPX-AP-\*-EP-\*) 
 
+### Fixed
+- Updated `requests` dependency to remove missing `chardet` error. Will now only raise a warning. Can be fixed by manually installing `chardet`
+
 ## v0.5.1 - 21.06.24
 ### Fixed
 - CPX-AP: Limited the modbus timeout to a minimum 100 ms. Values lower that that could lead to exclusion from the system.